IOS-7560: Simplify repository and improve naming
